Add helpers for Alias and RealName
Some systems are outputting a blank for Alias, so this adds new AliasOrRealName variable that will default to the Alias, but if it's blank, fallback to the Real Name.
A related RealNameOrAlias variable is also added for the reverse fallback scenario, in case a programmer prefers working with real names.